[FlyDSL] move get_element_ptr out of buffer_ops
It is a raw LLVM GEP builder, not a buffer op, and its 14 callers all want a
plain pointer rather than a descriptor. It joins create_llvm_ptr in
kernels_common, which is where the raw-pointer helpers now live.

buffer_ops needed it for one internal byte GEP in from_memref; that is inlined
rather than imported back, since kernels_common imports buffer_ops and the
dependency must not become a cycle. The moved copy also drops its own reach
into buffer_ops._unwrap_value, and uses kernels_common's T, where T.i8 is
already a type rather than a factory.

buffer_ops is now just buffer load/store and resource creation:
  __all__ = [buffer_load, buffer_store, create_buffer_resource,
             create_buffer_resource_from_addr]

Validated: MoE aux 420/420 zero errors, gather-reduce 216/216, causal_conv1d
11/11, qk_norm 18/18 err=0, the EP dispatch/combine and mega_moe stage kernels
still compile, and ISA unchanged -- 19 gfx1250 scatter modules and all 3 fmha
modules byte-identical.
