+++ aiter/ops/flydsl/kernels/mxfp4_gemm_common.py
@@ -10,6 +10,7 @@ from flydsl.expr.typing import T
 from aiter.ops.flydsl.kernels import buffer_ops
 
 from . import dpp_utils
+from .kernels_common import get_element_ptr
 
 kStages = 2
 kBS_stride_k0_dw = 64
@@ -36,9 +37,7 @@ def _lds_base_ptr3(lds_view):
 
 
 def _gep3(base_ptr, byte_off_i32):
-    return buffer_ops.get_element_ptr(
-        base_ptr, byte_offset=byte_off_i32, elem_type=T.i8
-    )
+    return get_element_ptr(base_ptr, byte_offset=byte_off_i32, elem_type=T.i8)
 
 
 def _global_base_ptr1(addr_i64):
@@ -47,9 +46,7 @@ def _global_base_ptr1(addr_i64):
 
 
 def _gep1(base_ptr, byte_off_i32):
-    return buffer_ops.get_element_ptr(
-        base_ptr, byte_offset=byte_off_i32, elem_type=T.i8
-    )
+    return get_element_ptr(base_ptr, byte_offset=byte_off_i32, elem_type=T.i8)
 
 
 def _global_ptr1(arg, byte_off_i32):
